fix: moved migrations slice to a migrate package
